pull skymodel related code out of pyuvsim
