+++ pyuvsim/simsetup.py
@@ -98,262 +98,6 @@ def _config_str_to_dict(config_str):
 
     return param_dict
 
-
-def skymodel_to_array(sky):
-    """
-    Return a recarray of source components from a given SkyModel object.
-    """
-
-    fieldtypes = ['U10', 'f8', 'f8', 'f8', 'f8']
-    fieldnames = ['source_id', 'ra_j2000', 'dec_j2000', 'flux_density_I', 'frequency']
-    fieldshapes = [()] * 3 + [(sky.Nfreqs,)] * 2
-
-    dt = np.dtype(list(zip(fieldnames, fieldtypes, fieldshapes)))
-    if sky.Nfreqs == 1:
-        sky.freq_array = sky.freq_array[:, None]
-
-    arr = np.empty(sky.Ncomponents, dtype=dt)
-    arr['source_id'] = sky.name
-    arr['ra_j2000'] = sky.ra.deg
-    arr['dec_j2000'] = sky.dec.deg
-    arr['flux_density_I'] = sky.stokes[0, :, :].T   # Swaps component and frequency axes
-    arr['frequency'] = sky.freq_array
-
-    return arr
-
-
-def array_to_skymodel(catalog_table):
-    """
-    Make a SkyModel object from a recarray.
-    """
-
-    ra = Angle(catalog_table['ra_j2000'], units.deg)
-    dec = Angle(catalog_table['dec_j2000'], units.deg)
-    ids = catalog_table['source_id']
-    flux_I = np.atleast_1d(catalog_table['flux_density_I'])
-    if flux_I.ndim == 1:
-        flux_I = flux_I[:, None]
-    stokes = np.pad(np.expand_dims(flux_I, 2), ((0, 0), (0, 0), (0, 3)), 'constant').T
-    rise_lst = None
-    set_lst = None
-    source_freqs = np.atleast_1d(catalog_table['frequency'][0])
-
-    if 'rise_lst' in catalog_table.dtype.names:
-        rise_lst = catalog_table['rise_lst']
-        set_lst = catalog_table['set_lst']
-
-    sourcelist = SkyModel(ids, ra, dec, stokes, Nfreqs=source_freqs.size,
-                          freq_array=source_freqs, rise_lst=rise_lst, set_lst=set_lst)
-
-    return sourcelist
-
-
-def source_cuts(catalog_table, input_uv=None, latitude_deg=None, horizon_buffer=0.04364,
-                min_flux=None, max_flux=None):
-    """
-    Performing flux and horizon selections on recarray of source components.
-
-    Args:
-        catalog_table: recarray of source catalog information. Must have the columns:
-                       'source_id', 'ra_j2000', 'dec_j2000', 'flux_density_I', 'frequency'
-        input_uv: (UVData)
-            If not None, can set lst_array and latitude_deg if either or both is missing.
-        latitude_deg: Latitude of telescope in degrees. Used to estimate rise/set lst.
-        horizon_buffer: Angle (float, in radians)
-            Buffer for coarse horizon cut. Default is about 10 minutes of sky rotation. `SkyModel`
-            components whose calculated altitude is less than `horizon_buffer` are excluded.
-            Caution! The altitude calculation does not account for precession/nutation of the Earth.
-            The buffer angle is needed to ensure that the horizon cut doesn't exclude sources near
-            but above the horizon. Since the cutoff is done using lst, and the lsts are calculated
-            with astropy, the required buffer should _not_ drift with time since the J2000 epoch.
-            The default buffer has been tested around julian date 2457458.0.
-        min_flux: Minimum stokes I flux to select [Jy]
-        max_flux: Maximum stokes I flux to select [Jy]
-    Returns:
-        A new recarray of source components, with additional columns for rise and set lst.
-    """
-
-    # Option to pass a UVData object to get telescope-specific parameters
-    if input_uv:
-        if latitude_deg is None:
-            latitude_deg = input_uv.telescope_location_lat_lon_alt_degrees[0]
-
-    coarse_horizon_cut = latitude_deg is not None
-
-    if coarse_horizon_cut:
-        lat_rad = np.radians(latitude_deg)
-        buff = horizon_buffer
-
-    if min_flux:
-        catalog_table = catalog_table[catalog_table['flux_density_I'] > min_flux]
-
-    if max_flux:
-        catalog_table = catalog_table[catalog_table['flux_density_I'] < max_flux]
-
-    ra = Angle(catalog_table['ra_j2000'], units.deg)
-    dec = Angle(catalog_table['dec_j2000'], units.deg)
-
-    if coarse_horizon_cut:
-        tans = np.tan(lat_rad) * np.tan(dec.rad)
-        nonrising = tans < -1
-
-        catalog_table = catalog_table[~nonrising]
-        tans = tans[~nonrising]
-
-        ra = ra[~nonrising]
-
-    if coarse_horizon_cut:
-        with warnings.catch_warnings():
-            warnings.filterwarnings('ignore', message='invalid value encountered',
-                                    category=RuntimeWarning)
-            rise_lst = ra.rad - np.arccos((-1) * tans) - buff
-            set_lst = ra.rad + np.arccos((-1) * tans) + buff
-
-            rise_lst[rise_lst < 0] += 2 * np.pi
-            set_lst[set_lst < 0] += 2 * np.pi
-            rise_lst[rise_lst > 2 * np.pi] -= 2 * np.pi
-            set_lst[set_lst > 2 * np.pi] -= 2 * np.pi
-
-        catalog_table = recfunctions.append_fields(catalog_table, ['rise_lst', 'set_lst'],
-                                                   [rise_lst, set_lst],
-                                                   usemask=False)
-
-    return catalog_table
-
-
-def read_votable_catalog(gleam_votable, input_uv=None, source_select_kwds={}, return_table=False):
-    """
-    Creates a list of pyuvsim source objects from a votable catalog.
-
-    Tested on: GLEAM EGC catalog, version 2
-
-    Args:
-        gleam_votable: str
-            Path to votable catalog file.
-        input_uv: :class:`~pyuvdata.UVData` object
-            The UVData object for the simulation (needed for horizon cuts)
-        return_table: bool, optional
-            Whether to return the astropy table instead of a list of Source objects.
-        source_select_kwds: dict, optional
-            Dictionary of keywords for source selection Valid options:
-
-            * `lst_array`: For coarse RA horizon cuts, lsts used in the simulation [radians]
-            * `latitude_deg`: Latitude of telescope in degrees. Used for declination coarse
-               horizon cut.
-            * `horizon_buffer`: Angle (float, in radians) of buffer for coarse horizon cut.
-              Default is about 10 minutes of sky rotation. (See caveats in
-              :func:`array_to_skymodel` docstring)
-            * `min_flux`: Minimum stokes I flux to select [Jy]
-            * `max_flux`: Maximum stokes I flux to select [Jy]
-
-    Returns:
-        if return_table, recarray of source parameters, otherwise :class:`pyuvsim.SkyModel` instance
-    """
-
-    class Found(Exception):
-        pass
-
-    resources = parse(gleam_votable).resources
-    try:
-        for rs in resources:
-            for tab in rs.tables:
-                if 'GLEAM' in tab.array.dtype.names:
-                    raise Found
-    except Found:
-        table = tab.to_table()  # Convert to astropy Table
-
-    fieldnames = ['GLEAM', 'RAJ2000', 'DEJ2000', 'Fintwide']
-    newnames = ['source_id', 'ra_j2000', 'dec_j2000', 'flux_density_I', 'frequency']
-    data = table[fieldnames]
-    Nsrcs = len(data)
-    freq = 200e6
-    for t in data.colnames:
-        i = fieldnames.index(t)
-        data[t] = data[t]
-        data[t].name = newnames[i]
-    data.add_column(table.Column(np.ones(Nsrcs) * freq, name='frequency'))
-    data = data.as_array().data
-
-    if len(source_select_kwds) > 0:
-        data = source_cuts(data, input_uv=input_uv, **source_select_kwds)
-
-    if return_table:
-        return data
-
-    return array_to_skymodel(data)
-
-
-def read_text_catalog(catalog_csv, input_uv=None, source_select_kwds={}, return_table=False):
-    """
-    Read in a text file of sources.
-
-    Args:
-        catalog_csv: str
-            Path to tab separated value file with the following expected columns
-            (For now, all sources are flat spectrum):
-
-            *  `Source_ID`: source name as a string of maximum 10 characters
-            *  `ra_j2000`: right ascension at J2000 epoch, in decimal degrees
-            *  `dec_j2000`: declination at J2000 epoch, in decimal degrees
-            *  `flux_density_I`: Stokes I flux density in Janskys
-            *  `frequency`: reference frequency (for future spectral indexing) [Hz]
-
-        input_uv: :class:`pyuvdata.UVData` object
-            The UVData object for the simulation (needed for horizon cuts)
-        source_select_kwds: dict, optional
-            Dictionary of keywords for source selection. Valid options:
-
-            * `lst_array`: For coarse RA horizon cuts, lsts used in the simulation [radians]
-            * `latitude_deg`: Latitude of telescope in degrees. Used for declination coarse
-            *  horizon cut.
-            * `horizon_buffer`: Angle (float, in radians) of buffer for coarse horizon cut.
-              Default is about 10 minutes of sky rotation. (See caveats in
-              :func:`array_to_skymodel` docstring)
-            * `min_flux`: Minimum stokes I flux to select [Jy]
-            * `max_flux`: Maximum stokes I flux to select [Jy]
-
-    Returns:
-        :class:`pyuvsim.SkyModel`
-    """
-    with open(catalog_csv, 'r') as cfile:
-        header = cfile.readline()
-    header = [h.strip() for h in header.split() if not h[0] == '[']  # Ignore units in header
-    dt = np.format_parser(['U10', 'f8', 'f8', 'f8', 'f8'],
-                          ['source_id', 'ra_j2000', 'dec_j2000', 'flux_density_I', 'frequency'],
-                          header)
-
-    catalog_table = np.genfromtxt(catalog_csv, autostrip=True, skip_header=1,
-                                  dtype=dt.dtype)
-
-    catalog_table = np.atleast_1d(catalog_table)
-
-    if len(source_select_kwds) > 0:
-        catalog_table = source_cuts(catalog_table, input_uv=input_uv, **source_select_kwds)
-
-    if return_table:
-        return catalog_table
-
-    return array_to_skymodel(catalog_table)
-
-
-def write_catalog_to_file(filename, catalog):
-    """
-    Writes out a catalog to a text file, readable with simsetup.read_catalog_text()
-
-    Args:
-        filename: Path to output file (string)
-        catalog: pyuvsim.SkyModel object
-    """
-    with open(filename, 'w+') as fo:
-        fo.write("SOURCE_ID\tRA_J2000 [deg]\tDec_J2000 [deg]\tFlux [Jy]\tFrequency [Hz]\n")
-        arr = skymodel_to_array(catalog)
-        for src in arr:
-            srcid, ra, dec, flux_i, freq = src
-
-            fo.write("{}\t{:f}\t{:f}\t{:0.2f}\t{:0.2f}\n".format(
-                srcid, ra, dec, flux_i[0], freq[0]))
-
-
 def create_mock_catalog(time, arrangement='zenith', array_location=None, Nsrcs=None,
                         alt=None, save=False, min_alt=None, rseed=None, return_table=False):
     """
